requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.5
django-bootstrap-datepicker-plus==5.0.3
django-filter==22.1
django-heroku==0.3.1
django-import-export==3.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.4.0
psycopg2==2.9.5
pydantic==1.10.4
pytz==2022.7
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tablib==3.3.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
